refactor(ui): extract the pause menu's world dim into a shared module
Move the VR comfort dim (layer, distance, gaze-pose resolution, tuning
constants) from pause_menu.rs into ui/world_dim.rs unchanged, with the
debug provenance tag as a parameter, so the upcoming cyber-interface use
mode can dim the world behind its panel with the same code instead of a
copy. Pause menu behavior is unchanged; its dim tests still exercise the
shared functions.
